Commit nästan färdig
Feature:// added achivements so you can see your progress. just now there is only one achivements due to this not being the completed version yet.
you get an achivement for making a acount

Styling:// gave the app a needed makeover so it looks a little better. using blue and white as my primary color.

So far the only things the app is missing is some pictures and some text that are placeholder other than that the app works as intended. And of course some achivements are still missing but those will be added later on. There are nearly finished.